backend/requirements.txt
# Web API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Storage dependencies
boto3==1.29.7
python-dotenv==1.0.0

# Existing dependencies from base
mediapipe==0.10.7
opencv-python-headless==4.8.1.78  # Use headless for server deployment
pandas==2.1.3
numpy==1.24.3
Pillow==10.1.0  # For image processing